requirements.txt
Flask==3.1.2
gunicorn==23.0.0
google-cloud-pubsub==2.34.0